Code compiles but has not been tested. Need to check what else can be unified to avoid duplicate code


 Conflicts:
	sensing/autoware_pointcloud_preprocessor/CMakeLists.txt
	sensing/autoware_pointcloud_preprocessor/config/concatenate_and_time_sync_node.param.yaml
	sensing/autoware_pointcloud_preprocessor/docs/concatenate-data.md
	sensing/autoware_pointcloud_preprocessor/include/autoware/pointcloud_preprocessor/concatenate_data/cloud_collector.hpp
	sensing/autoware_pointcloud_preprocessor/include/autoware/pointcloud_preprocessor/concatenate_data/combine_cloud_handler.hpp
	sensing/autoware_pointcloud_preprocessor/include/autoware/pointcloud_preprocessor/concatenate_data/concatenate_and_time_sync_node.hpp
	sensing/autoware_pointcloud_preprocessor/schema/concatenate_and_time_sync_node.schema.json
	sensing/autoware_pointcloud_preprocessor/src/concatenate_data/cloud_collector.cpp
	sensing/autoware_pointcloud_preprocessor/src/concatenate_data/combine_cloud_handler.cpp
	sensing/autoware_pointcloud_preprocessor/src/concatenate_data/concatenate_and_time_sync_node.cpp
	sensing/autoware_pointcloud_preprocessor/test/test_concatenate_node_component.py
	sensing/autoware_pointcloud_preprocessor/test/test_concatenate_node_unit.cpp

+++ planning/autoware_freespace_planning_algorithms/src/astar_search.cpp
@@ -388,7 +388,7 @@ double AstarSearch::getExpansionDistance(const AstarNode & current_node) const
 double AstarSearch::getSteeringCost(const int steering_index) const
 {
   return planner_common_param_.curve_weight *
-         (abs(steering_index) / planner_common_param_.turning_steps);
+         (static_cast<double>(abs(steering_index)) / planner_common_param_.turning_steps);
 }
 
 double AstarSearch::getSteeringChangeCost(
